requirements.txt
jsonschema==3.2.0
PyYAML==5.4.1